requirements.txt
telethon==1.28.5
python-telegram-bot==13.15
python-dotenv==1.0.0
